[CO-354] Introduce `NetworkMagic`
We add the `NetworkMagic` datatype, which is isomorphic to `Maybe Word8`
but provides more context & type-safety.

For ease of configuration, we also introduce the `RequiresNetworkMagic`
helper datatype, which actually goes into the configuration file,
providing a richer descriptor than a Bool (with functional equivalence).

In order to keep this commit (relatively) small, I mostly ignore the
new `NetworkMagic` argument which is now provided by the updated
config functions. In subsequent commits we can pipe the `NetworkMagic`
down to where we will need it - Address verification and generation.