[CO-354] Introduce `NetworkMagic`, routing through `ProtocolMagic`
We add the `NetworkMagic` datatype, which is isomorphic to `Maybe Word8`
but provides more context & type-safety.

For ease of configuration, we also introduce the `RequiresNetworkMagic`
helper datatype, which actually goes into the configuration file,
providing a richer descriptor than a Bool (with functional equivalence).

Unlike our first approach, we route `RequiresNetworkMagic` through the
`ProtocolMagic` datatype, in order to minimize the number of LOCs
changed (limiting our "error surface", as this is a relatively complex
PR).

! merged with "[CO-354] WIP: rest of changes for PMPrime"
! merged with "[CO-354] Push modified `ProtocolMagic` through codebase"

All src/ and test/ files build. Several testsuites fail. Some tests
fail simply because of the `P.error` calls I have made in cases
where we have an undefined `requiresNetworkMagic` in a `ProtocolMagic`
(which, due to strict fields), gets forced, even though it should be
unused. Other tests fail for what seem to be legitimate reasons.

I will look into those failures, and I have an idea for how to get
around the `requiresNetworkMagic` issue.

! merged with "[CO-354] Fix `core` & `chain` testsuites"