[CO-354] Fix some `generator` tests
Tests were failing because the Utxo generation was reading from
a static `CoreConfiguration` (via HasCoreConfigurations typeclass)
which was not lined up with the dynamic ProtocolMagic & NetworkMagic
that was being passed to the rest of the test code.

The configuration pipeline is very confusing at this point, and I
think we should try to simplify it.